[ISSUE-19] Refactoring ASM based transformer accordingly
