+++ impl/asm-based-transformer/src/main/java/org/wildfly/transformer/asm/TransformerBuilderImpl.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.wildfly.transformer.nodeps;
+package org.wildfly.transformer.asm;
 
 import static java.lang.Thread.currentThread;
